-feat 파일 접근시 권한 체크, pdf/mxl 업로드 api 분리/ 한번에 한개씩 업로드하게 변경

+++ src/main/java/com/ai/FlatServer/file/controller/FileController.java
@@ -4,7 +4,12 @@ import com.ai.FlatServer.file.dto.request.FilePatchRequest;
 import com.ai.FlatServer.file.dto.request.PdfUploadRequest;
 import com.ai.FlatServer.file.dto.response.FileDto;
 import com.ai.FlatServer.file.dto.response.FileNameInfo;
+import com.ai.FlatServer.file.respository.dao.FileInfo;
 import com.ai.FlatServer.file.service.FileService;
+import com.ai.FlatServer.folder.service.FolderService;
+import com.ai.FlatServer.rabbitmq.service.MessageService;
+import com.ai.FlatServer.user.repository.entity.User;
+import com.ai.FlatServer.user.service.UserService;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.util.ArrayList;
@@ -34,63 +39,68 @@ import org.springframework.web.multipart.MultipartFile;
 public class FileController {
 
     private final FileService fileService;
+    private final FolderService folderService;
+    private final UserService userService;
+    private final MessageService messageService;
 
     @GetMapping("/mxl/{fileId}")
-    public ResponseEntity<UrlResource> getMxl(@PathVariable Long fileId) {
-        try {
-            FileDto xmlFileDto = fileService.getMxl(fileId);
-            return ResponseEntity.ok()
-                    .header(HttpHeaders.CONTENT_DISPOSITION, ContentDisposition.attachment()
-                            .filename(xmlFileDto.getEncodedFileName())
-                            .build()
-                            .toString())
-                    .contentType(MediaType.APPLICATION_OCTET_STREAM)
-                    .body(xmlFileDto.getFile());
-        } catch (MalformedURLException e) {
-            return ResponseEntity.internalServerError().build();
-        }
+    public ResponseEntity<UrlResource> getMxl(@PathVariable Long fileId) throws MalformedURLException {
+        FileDto xmlFileDto = fileService.getMxl(fileId);
+        return ResponseEntity.ok()
+                .header(HttpHeaders.CONTENT_DISPOSITION, ContentDisposition.attachment()
+                        .filename(xmlFileDto.getEncodedFileName())
+                        .build()
+                        .toString())
+                .contentType(MediaType.APPLICATION_OCTET_STREAM)
+                .body(xmlFileDto.getFile());
     }
 
 
     @GetMapping("/pdf/{fileUid}")
-    public ResponseEntity<UrlResource> getPdf(@PathVariable String fileUid) {
-        try {
-            FileDto fileDto = fileService.getPdf(fileUid);
-            return ResponseEntity.ok()
-                    .header(HttpHeaders.CONTENT_DISPOSITION, ContentDisposition.attachment()
-                            .filename(fileDto.getEncodedFileName() + ".pdf")
-                            .build()
-                            .toString())
-                    .contentType(MediaType.APPLICATION_PDF)
-                    .body(fileDto.getFile());
-        } catch (MalformedURLException e) {
-            return ResponseEntity.internalServerError().build();
-        }
+    public ResponseEntity<UrlResource> getPdf(@PathVariable String fileUid) throws MalformedURLException {
+        FileInfo fileInfo = fileService.getPdf(fileUid);
+        userService.checkFileAuthority(fileInfo);
+        FileDto fileDto = fileService.encodePdf(fileInfo);
+        return ResponseEntity.ok()
+                .header(HttpHeaders.CONTENT_DISPOSITION, ContentDisposition.attachment()
+                        .filename(fileDto.getEncodedFileName() + ".pdf")
+                        .build()
+                        .toString())
+                .contentType(MediaType.APPLICATION_PDF)
+                .body(fileDto.getFile());
     }
 
     @GetMapping("/fav")
     public ResponseEntity<List<FileNameInfo>> getFavs() {
-        return ResponseEntity.ok(fileService.getFavs());
+        User user = userService.getCurrentUser();
+        return ResponseEntity.ok(fileService.getFavs(user));
     }
 
 
-    @PostMapping
+    @PostMapping("/pdf")
     public ResponseEntity<ArrayList<Long>> uploadPdf(
-            @RequestPart(value = "dto", required = false) PdfUploadRequest pdfUploadRequest,
-            @RequestPart(value = "files") List<MultipartFile> multipartFile) {
-        ArrayList<Long> Ids = new ArrayList<>();
-        try {
-            for (MultipartFile m : multipartFile) {
-                Ids.add(fileService.saveFile(m, pdfUploadRequest));
-            }
-        } catch (IOException e) {
-            return ResponseEntity.internalServerError().build();
-        }
-        return ResponseEntity.status(201).body(Ids);
+            @RequestPart(value = "dto") PdfUploadRequest pdfUploadRequest,
+            @RequestPart(value = "file") MultipartFile multipartFile) throws IOException {
+        User user = userService.getCurrentUser();
+        folderService.checkFolderAuthority(user, pdfUploadRequest.getFolderId());
+
+        String s = fileService.savePdf(multipartFile, pdfUploadRequest);
+        messageService.sendTransformRequestMessage(s);
+        return ResponseEntity.status(201).build();
+    }
+
+    @PostMapping("/mxl")
+    public ResponseEntity<Boolean> uploadMxl(
+            @RequestPart(value = "file") MultipartFile multipartFile) throws IOException {
+        fileService.saveMxl(multipartFile);
+        return ResponseEntity.status(201).build();
     }
 
     @DeleteMapping("/{targetFileId}")
     public ResponseEntity<Boolean> deleteFile(@PathVariable Long targetFileId) {
+        User user = userService.getCurrentUser();
+        fileService.checkFileAuthority(user, targetFileId);
+
         fileService.removeFile(targetFileId);
         return ResponseEntity.ok().build();
     }
@@ -99,6 +109,9 @@ public class FileController {
     @PatchMapping("/{targetFileId}")
     public ResponseEntity<Boolean> patchFile(@PathVariable Long targetFileId,
                                              @RequestBody FilePatchRequest filePatchRequest) {
+        User user = userService.getCurrentUser();
+        fileService.checkFileAuthority(user, targetFileId);
+
         fileService.patchFile(targetFileId, filePatchRequest);
         return ResponseEntity.ok().build();
     }
